+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT RELEASE
-* Use a Semaphore to limit concurrency of hashtag search API requests; add Config.maxConcurrentTootRequests
+* HashtagParticipationScorer uses the square root of the number of participations to tamp down runaway scores
+* Use a Semaphore to limit concurrency of hashtag search API requests; add `Config.maxConcurrentTootRequests`
 * Exclude followed tags from the pulling of hashtag participation toots (followed tags are already in timeline)
 * Refactor common handling for cacheable toot fetches